fix test_ordering_satisfy_with_equivalence2
